# Implemented KMP search optimizations to previously implemented Trie algorithm