--- a/Detection/src/Extensions/IPrefixTrie.cs
+++ b/Detection/src/Extensions/IPrefixTrie.cs
@@ -1,0 +1,10 @@
+// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
+
+namespace Wangkanai.Detection.Extensions;
+
+public interface IPrefixTrie
+{
+	bool ContainsWithAnyIn(ReadOnlySpan<char> source);
+	bool StartsWithAnyIn(ReadOnlySpan<char>   source);
+	bool IsEnd { get; }
+}