--- a/Detection/tests/Extensions/KmpPrefixTrieTests.cs
+++ b/Detection/tests/Extensions/KmpPrefixTrieTests.cs
@@ -1,0 +1,59 @@
+﻿// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
+
+namespace Wangkanai.Detection.Extensions;
+
+public class KmpPrefixTrieTests: BasePrefixTrieTests<KmpPrefixTrie>
+{
+	protected override KmpPrefixTrie Create(string[]? keywords)
+	{
+		return new KmpPrefixTrie(keywords);
+	}
+
+	[Fact]
+	public void KeywordsIsNull_Test()
+	{
+		KeywordsIsNull();
+	}
+
+	[Fact]
+	public void KeywordsIsEmpty_Test()
+	{
+		KeywordsIsEmpty();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceIsNull_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceIsNull_SeedIsZero();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceIsEmpty_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceIsEmpty_SeedIsZero();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceIsWhiteSpace_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceIsWhiteSpace_SeedIsZero();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceExist_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceExist_SeedIsZero();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceSubset_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceSubset_SeedIsZero();
+	}
+
+	[Fact]
+	public void KeywordsExist_SourceNotSub_SeedIsZero_Test()
+	{
+		KeywordsExist_SourceNotSub_SeedIsZero();
+	}
+}