… Detection/src/Extensions/IndexTree.cs
… Detection/tests/Extensions/IndexTreeTests.cs
… Detection/tests/Extensions/KmpPrefixTrieEdgeCaseTests.cs
… Detection/tests/Extensions/PrefixTrieEdgeCaseTests.cs