Implemented separate KMP search optimization to previously implemented Prefix Trie.

~17% improvement for device E2E benchmark
~4% improvement for browser E2E benchmark
~3% improvement for engine E2E benchmark

Majority of these improvements are done by introducing KMP algorithm optimizations to the Trie search algorithm, for longer browser agent strings has greater performance benefit and improvement.

*Before:*
Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
Device | 15.090 μs | 0.0919 μs | 0.0815 μs | 2.5177 | 0.0153 | 41.25 KB
Platform | 9.571 μs | 0.1021 μs | 0.0905 μs | 2.5330 | 0.0153 | 41.63 KB
Engine | 39.124 μs | 0.6082 μs | 0.5079 μs | 8.7891 | - | 144.4 KB
Browser | 72.928 μs | 0.8599 μs | 0.7623 μs | 15.5029 | - | 254.16 KB

*After:*
Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
Device | 12.649 μs | 0.1200 μs | 0.1064 μs | 2.5177 | 0.0153 | 41.25 KB
Platform | 9.664 μs | 0.1038 μs | 0.0867 μs | 2.5330 | 0.0153 | 41.63 KB
Engine | 37.763 μs | 0.3858 μs | 0.3420 μs | 8.7891 | - | 144.4 KB
Browser | 69.916 μs | 0.5773 μs | 0.5400 μs | 15.5029 | - | 254.16 KB

*Setup:*
BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.3476)
AMD Ryzen 9 9950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 9.0.200
  [Host]     : .NET 8.0.14 (8.0.1425.11118), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.14 (8.0.1425.11118), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI