@wangkanai 2025-03-28
This pull request introduces significant changes to the `Detection` project, focusing on replacing the `IndexTree` structure with a more efficient `IPrefixTrie` interface and its implementations. The changes include creating new benchmark tests, updating existing collections and services, and adding new classes to support the new trie structure.

### Key Changes:

#### New Benchmark Tests:
* Added `TrieBenchmark` class to benchmark the performance of `PrefixTrie` and `KmpPrefixTrie` implementations with various keyword sets and text sizes (`Detection/benchmark/TrieBenchmark.cs`).

#### Interface and Implementations:
* Introduced `IPrefixTrie` interface to define the contract for trie implementations (`Detection/src/Extensions/IPrefixTrie.cs`).
* Added `KmpPrefixTrie` class implementing `IPrefixTrie` with optimizations using the Knuth-Morris-Pratt algorithm (`Detection/src/Extensions/KmpPrefixTrie.cs`).
* Added `PrefixTrie` class implementing `IPrefixTrie` as a basic trie structure (`Detection/src/Extensions/PrefixTrie.cs`).

#### Updates to Collections:
* Updated `MobileCollection` and `TabletCollection` to use `IPrefixTrie` instead of `IndexTree` for keyword searches (`Detection/src/Collections/MobileCollection.cs`, `Detection/src/Collections/TabletCollection.cs`). [[1]](diffhunk://#diff-30610a047c46a94eae4d6da1a9bdd329b8374ebb10b42c5066c58f7bb8304399L42-R42) [[2]](diffhunk://#diff-30610a047c46a94eae4d6da1a9bdd329b8374ebb10b42c5066c58f7bb8304399L137-R137) [[3]](diffhunk://#diff-12022d3bf4a4f03d5875c551d008a54f0047b495b757dcbcbcb7b0ff8f006a99L20-R20)

#### Service Modifications:
* Modified various service methods to utilize the new `IPrefixTrie` interface for searching device and platform information (`Detection/src/Services/CrawlerService.cs`, `Detection/src/Services/DeviceService.cs`, `Detection/src/Services/PlatformService.cs`). [[1]](diffhunk://#diff-bb4e3f02f695733763718972464f8f68c9c28d0ca6561e7b7300ef25e39c8f90L66-R75) [[2]](diffhunk://#diff-606a5aff318cb2dd18b2e30dd7894b7a46cad6fb909068343eee1678c10a6bcbL47-R53) [[3]](diffhunk://#diff-63731daa6a233e501fccf2b003815af505d0aaf064c7df36cb00aaef7df95275L90-R102) [[4]](diffhunk://#diff-63731daa6a233e501fccf2b003815af505d0aaf064c7df36cb00aaef7df95275L142-R153)

These changes aim to improve the efficiency and performance of keyword searches within the `Detection` project by leveraging the new trie structures.