@wangkanai 2025-03-28
@multiarc 

I am very impressed by your contribution. This is an efficient way of performing this search pattern, and it's the first time that I have encountered this method. 

Thank you very much.