@wangkanai 2025-03-28
@multiarc https://github.com/wangkanai/wangkanai/pull/1036 finishing up the code.

I'm also reading up more on Knuth–Morris–Pratt algorithm.