@multiarc 2025-03-28
@wangkanai direct application of Knuth–Morris–Pratt is just string search, I adapter principles of this algorithm and core features into Prefix Trie. Primary complexity is to walk all possible trie paths and creating individual lookback tables and references to optimize overall search. So there are all benefits of Prefix Trie and its performance, but we have this lookback KMP table to continue search at a place closer to a string end and a little ahead in Prefix Trie - basically optimize comparative operations. Even though I used class instead of struct it is still faster and struct would actually be slower in this case because of lookbacks as you would have to use struct references which in return would be boxes and so actually slower. Maybe this is possible to optimize it further by using structs without copying and boxing it can be faster, but I never able to make this happen. When any sort of references involved classes are just superior.