@multiarc 2025-03-28
Also, both (Trie and KMP Trie) are not great in terms of memory usage when non ASCII characters are used with ASCII characters when distance between values are quite high in terms of Unicode character space. So, if you simply use any single language character space would be adequate, but multi-lang strings would consume more memory due to distance. Just natural caveat of Prefix Trie search when no hash is being calculated to limit space and range.